fix: merge bug fixes

+++ carsApp/back-end/back-end/src/main/java/sbnz/repository/BuyRequestRepository.java
@@ -3,5 +3,5 @@ package sbnz.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import sbnz.domain.BuyRequest;
 
-public interface BuyRequestRepository extends JpaRepository<BuyRequest, Long> {
+public interface BuyRequestRepository extends JpaRepository<BuyRequest, Integer> {
 }
